pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brain-box"
version = "1.0.0"
description = "Telegram-based AI knowledge management system"
requires-python = ">=3.10"
dependencies = [
    "python-telegram-bot[ext]>=21.0",
    "openai>=1.0",
    "python-dotenv>=1.0",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.19",
    "httpx[socks]>=0.27",
    "pdfplumber>=0.10",
    "python-docx>=1.0",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "PyGithub>=2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "black>=24.0",
    "ruff>=0.3",
    "pre-commit>=3.6",
]

[project.scripts]
brain-box = "app.main:main"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
include = ["app*"]
